Added checks to generator.

+++ include/common.h
@@ -27,7 +27,7 @@ extern "C" {
 static inline
 int32_t log10_pow2(int32_t n) {
   int64_t const log10_2 = 1292913986;
-  return n >= 0 ? (int32_t) (log10_2 * n >> 32) : -log10_pow2(-n) - 1;
+  return n >= 0 ? log10_2 * n >> 32 : -log10_pow2(-n) - 1;
 }
 
 #ifdef __cplusplus
